requirements.txt
discord.py == 1.3.3
pyyaml == 5.3.1
requests == 2.23.0
pony == 0.7.13